enet: packet PLAYER_MOVE - wip

+++ core/player.h
@@ -80,7 +80,7 @@ class Player : public InventoryElement
     InventoryElement * get_item_by_uid(size_t id);
     int get_id();
     // Player(int id);
-    Player(int id, SerializableCString && name, ItemLocation location, int thirst, int hunger, int nutrition);
+    Player(size_t id, SerializableCString&& name, ItemLocation location, int thirst, int hunger, int nutrition);
     int conversation(Player * who, Sentence * s, InventoryElement * el);
     void stop_conversation();
